Add files via upload
This code is a simulation of how in the restaurant problem (multi armed bandit problem), the values we calculate, i.e, the expected utility of a restaurant given we choose that resaturant gets closer and closer to the mean utility of that restaurant. This is done using a combination of sxloration and exploitation by the Epsilon-Greedy method where we have taken epsilon to be 0.1.